\name{Bt20}
\alias{Bt20}
\title{Effect of DFMO on breast cancer tumors}
\description{
  The data are the result of an experiment, that we will refer to as the
  Bt-20 experiment. The goal of the experiment was to determine whether
  hormone-independent human breast cancer cells growing in nude mice
  manifest sensitivity to the polyamine-biosynthetic inhibitor
  DFMO. Mice bearing established tumors to one of six groups: 0\%
  (control), 0.5\%, 1\%, 2\% and 3\%. Tumors in each group were measured
  at day 0, 3, 7, 10, 14 and 16.
}
\usage{
data(Bt20)
}
\format{ A matrix, the first column contains the group ids (which dose),
  the second column contains the animal ids. Columns 3 to 8 contains the
  measurements at the time points 0, 3, 7, 10, 14 and 16, respectively.
}
\source{Daniel F. Heitjan, Andrea Manni and Richard J. Santen.Statistical Analysis of in Vivo Tumor Growth Experiments.
  Cancer Research 1993.53:6042-6050. 
}
\references{
  Daniel F. Heitjan, Andrea Manni and Richard J. Santen.Statistical Analysis of in Vivo Tumor Growth Experiments.
  Cancer Research 1993.53:6042-6050.
}  
\keyword{datasets}
